Complete ES6 conversion.
