+++ saylua/static-source/es6/Main/Notifications.js
@@ -1,4 +1,4 @@
-import addClass from './Utils';
+import {addClass} from 'Utils';
 
 window.addEventListener('load', function () {
   var notificationCloseButtons = document.querySelectorAll('.notification a.close-button');
